+++ catkin_ws/src/10-lane-control/lane_filter/include/lane_filter/__init__.py
@@ -1,1 +1,1 @@
-# Empty, for now.
+from .lane_filter import *
